+++ saleor/graphql/page/tests/test_page.py
@@ -345,7 +345,7 @@ def test_page_create_mutation(staff_api_client, permission_manage_pages, page_ty
 
 
 @freeze_time("1914-06-28 10:50")
-@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_for_event.delay")
+@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_async")
 def test_page_create_trigger_page_webhook(
     mocked_webhook_trigger,
     staff_api_client,
@@ -384,7 +384,7 @@ def test_page_create_trigger_page_webhook(
     expected_data = generate_page_payload(page, staff_api_client.user)
 
     mocked_webhook_trigger.assert_called_once_with(
-        WebhookEventAsyncType.PAGE_CREATED, expected_data
+        expected_data, WebhookEventAsyncType.PAGE_CREATED
     )
 
 
@@ -1235,7 +1235,7 @@ def test_page_delete_mutation(staff_api_client, page, permission_manage_pages):
 
 
 @freeze_time("1914-06-28 10:50")
-@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_for_event.delay")
+@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_async")
 def test_page_delete_trigger_webhook(
     mocked_webhook_trigger, staff_api_client, page, permission_manage_pages, settings
 ):
@@ -1249,11 +1249,9 @@ def test_page_delete_trigger_webhook(
     assert data["page"]["title"] == page.title
     with pytest.raises(page._meta.model.DoesNotExist):
         page.refresh_from_db()
-
     expected_data = generate_page_payload(page, staff_api_client.user)
-
     mocked_webhook_trigger.assert_called_once_with(
-        WebhookEventAsyncType.PAGE_DELETED, expected_data
+        expected_data, WebhookEventAsyncType.PAGE_DELETED
     )
 
 
@@ -1395,7 +1393,7 @@ def test_update_page(staff_api_client, permission_manage_pages, page):
         assert attr_data in expected_attributes
 
 
-@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_for_event.delay")
+@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_async")
 @freeze_time("2020-03-18 12:00:00")
 def test_update_page_trigger_webhook(
     mocked_webhook_trigger, staff_api_client, permission_manage_pages, page, settings
@@ -1432,9 +1430,9 @@ def test_update_page_trigger_webhook(
     assert data["page"]["slug"] == new_slug
     page.publication_date = date(2020, 3, 18)
     expected_data = generate_page_payload(page, staff_api_client.user)
-
     mocked_webhook_trigger.assert_called_once_with(
-        WebhookEventAsyncType.PAGE_UPDATED, expected_data
+        expected_data,
+        WebhookEventAsyncType.PAGE_UPDATED,
     )