+++ saleor/plugins/webhook/tests/test_webhook.py
@@ -1,13 +1,17 @@
 import json
 from datetime import datetime
 from unittest import mock
+from unittest.mock import ANY, MagicMock
 from urllib.parse import urlencode
 
+import boto3
 import graphene
 import pytest
 from django.contrib.auth.tokens import default_token_generator
+from django.core.serializers import serialize
 from django.utils import timezone
 from freezegun import freeze_time
+from kombu.asynchronous.aws.sqs.connection import AsyncSQSConnection
 
 from .... import __version__
 from ....account.notifications import (
@@ -15,9 +19,10 @@ from ....account.notifications import (
     send_account_confirmation,
 )
 from ....app.models import App
+from ....core import EventDeliveryStatus
+from ....core.models import EventDelivery, EventDeliveryAttempt, EventPayload
 from ....core.notification.utils import get_site_context
 from ....core.notify_events import NotifyEventType
-from ....core.utils.json_serializer import CustomJsonEncoder
 from ....core.utils.url import prepare_url
 from ....discount.utils import fetch_catalogue_info
 from ....graphql.discount.mutations import convert_catalogue_info_to_global_ids
@@ -35,7 +40,11 @@ from ....webhook.payloads import (
     generate_sale_payload,
 )
 from ...manager import get_plugins_manager
-from ...webhook.tasks import trigger_webhooks_for_event
+from ...webhook.tasks import (
+    WebhookResponse,
+    send_webhook_request_async,
+    trigger_webhooks_async,
+)
 
 first_url = "http://www.example.com/first/"
 third_url = "http://www.example.com/third/"
@@ -54,7 +63,7 @@ third_url = "http://www.example.com/third/"
         (WebhookEventAsyncType.CUSTOMER_CREATED, 0, set()),
     ],
 )
-@mock.patch("saleor.plugins.webhook.tasks.send_webhook_request.delay")
+@mock.patch("saleor.plugins.webhook.tasks.send_webhook_request_async.delay")
 def test_trigger_webhooks_for_event_calls_expected_events(
     mock_request,
     event_name,
@@ -90,146 +99,149 @@ def test_trigger_webhooks_for_event_calls_expected_events(
         target_url="http://www.example.com/third/"
     )
     third_webhook.events.create(event_type=WebhookEventAsyncType.ANY)
-
-    trigger_webhooks_for_event(event_name, data="")
+    event_payload = EventPayload.objects.create()
+    trigger_webhooks_async(event_payload, event_name)
+    deliveries_called = {
+        EventDelivery.objects.get(id=delivery_id[0][0])
+        for delivery_id in mock_request.call_args_list
+    }
+    urls_called = {delivery.webhook.target_url for delivery in deliveries_called}
     assert mock_request.call_count == total_webhook_calls
-
-    target_url_calls = {call[0][2] for call in mock_request.call_args_list}
-    assert target_url_calls == expected_target_urls
+    assert urls_called == expected_target_urls
 
 
 @freeze_time("1914-06-28 10:50")
-@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_for_event.delay")
+@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_async")
 def test_order_created(mocked_webhook_trigger, settings, order_with_lines):
     settings.PLUGINS = ["saleor.plugins.webhook.plugin.WebhookPlugin"]
     manager = get_plugins_manager()
     manager.order_created(order_with_lines)
-
     expected_data = generate_order_payload(order_with_lines)
+
     mocked_webhook_trigger.assert_called_once_with(
-        WebhookEventAsyncType.ORDER_CREATED, expected_data
+        expected_data, WebhookEventAsyncType.ORDER_CREATED
     )
 
 
 @freeze_time("1914-06-28 10:50")
-@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_for_event.delay")
+@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_async")
 def test_order_confirmed(mocked_webhook_trigger, settings, order_with_lines):
     settings.PLUGINS = ["saleor.plugins.webhook.plugin.WebhookPlugin"]
     manager = get_plugins_manager()
     manager.order_confirmed(order_with_lines)
-
     expected_data = generate_order_payload(order_with_lines)
+
     mocked_webhook_trigger.assert_called_once_with(
-        WebhookEventAsyncType.ORDER_CONFIRMED, expected_data
+        expected_data, WebhookEventAsyncType.ORDER_CONFIRMED
     )
 
 
 @freeze_time("1914-06-28 10:50")
-@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_for_event.delay")
+@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_async")
 def test_draft_order_created(mocked_webhook_trigger, settings, order_with_lines):
     settings.PLUGINS = ["saleor.plugins.webhook.plugin.WebhookPlugin"]
     manager = get_plugins_manager()
     manager.draft_order_created(order_with_lines)
-
     expected_data = generate_order_payload(order_with_lines)
+
     mocked_webhook_trigger.assert_called_once_with(
-        WebhookEventAsyncType.DRAFT_ORDER_CREATED, expected_data
+        expected_data, WebhookEventAsyncType.DRAFT_ORDER_CREATED
     )
 
 
 @freeze_time("1914-06-28 10:50")
-@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_for_event.delay")
+@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_async")
 def test_draft_order_deleted(mocked_webhook_trigger, settings, order_with_lines):
     settings.PLUGINS = ["saleor.plugins.webhook.plugin.WebhookPlugin"]
     manager = get_plugins_manager()
     manager.draft_order_deleted(order_with_lines)
-
     expected_data = generate_order_payload(order_with_lines)
+
     mocked_webhook_trigger.assert_called_once_with(
-        WebhookEventAsyncType.DRAFT_ORDER_DELETED, expected_data
+        expected_data, WebhookEventAsyncType.DRAFT_ORDER_DELETED
     )
 
 
 @freeze_time("1914-06-28 10:50")
-@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_for_event.delay")
+@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_async")
 def test_draft_order_updated(mocked_webhook_trigger, settings, order_with_lines):
     settings.PLUGINS = ["saleor.plugins.webhook.plugin.WebhookPlugin"]
     manager = get_plugins_manager()
     manager.draft_order_updated(order_with_lines)
-
     expected_data = generate_order_payload(order_with_lines)
+
     mocked_webhook_trigger.assert_called_once_with(
-        WebhookEventAsyncType.DRAFT_ORDER_UPDATED, expected_data
+        expected_data, WebhookEventAsyncType.DRAFT_ORDER_UPDATED
     )
 
 
 @freeze_time("1914-06-28 10:50")
-@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_for_event.delay")
+@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_async")
 def test_customer_created(mocked_webhook_trigger, settings, customer_user):
     settings.PLUGINS = ["saleor.plugins.webhook.plugin.WebhookPlugin"]
     manager = get_plugins_manager()
     manager.customer_created(customer_user)
-
     expected_data = generate_customer_payload(customer_user)
+
     mocked_webhook_trigger.assert_called_once_with(
-        WebhookEventAsyncType.CUSTOMER_CREATED, expected_data
+        expected_data, WebhookEventAsyncType.CUSTOMER_CREATED
     )
 
 
 @freeze_time("1914-06-28 10:50")
-@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_for_event.delay")
+@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_async")
 def test_customer_updated(mocked_webhook_trigger, settings, customer_user):
     settings.PLUGINS = ["saleor.plugins.webhook.plugin.WebhookPlugin"]
     manager = get_plugins_manager()
     manager.customer_updated(customer_user)
-
     expected_data = generate_customer_payload(customer_user)
+
     mocked_webhook_trigger.assert_called_once_with(
-        WebhookEventAsyncType.CUSTOMER_UPDATED, expected_data
+        expected_data, WebhookEventAsyncType.CUSTOMER_UPDATED
     )
 
 
 @freeze_time("1914-06-28 10:50")
-@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_for_event.delay")
+@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_async")
 def test_order_fully_paid(mocked_webhook_trigger, settings, order_with_lines):
     settings.PLUGINS = ["saleor.plugins.webhook.plugin.WebhookPlugin"]
     manager = get_plugins_manager()
     manager.order_fully_paid(order_with_lines)
-
     expected_data = generate_order_payload(order_with_lines)
+
     mocked_webhook_trigger.assert_called_once_with(
-        WebhookEventAsyncType.ORDER_FULLY_PAID, expected_data
+        expected_data, WebhookEventAsyncType.ORDER_FULLY_PAID
     )
 
 
 @freeze_time("1914-06-28 10:50")
-@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_for_event.delay")
+@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_async")
 def test_product_created(mocked_webhook_trigger, settings, product):
     settings.PLUGINS = ["saleor.plugins.webhook.plugin.WebhookPlugin"]
     manager = get_plugins_manager()
     manager.product_created(product)
-
     expected_data = generate_product_payload(product)
+
     mocked_webhook_trigger.assert_called_once_with(
-        WebhookEventAsyncType.PRODUCT_CREATED, expected_data
+        expected_data, WebhookEventAsyncType.PRODUCT_CREATED
     )
 
 
 @freeze_time("1914-06-28 10:50")
-@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_for_event.delay")
+@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_async")
 def test_product_updated(mocked_webhook_trigger, settings, product):
     settings.PLUGINS = ["saleor.plugins.webhook.plugin.WebhookPlugin"]
     manager = get_plugins_manager()
     manager.product_updated(product)
-
     expected_data = generate_product_payload(product)
+
     mocked_webhook_trigger.assert_called_once_with(
-        WebhookEventAsyncType.PRODUCT_UPDATED, expected_data
+        expected_data, WebhookEventAsyncType.PRODUCT_UPDATED
     )
 
 
 @freeze_time("1914-06-28 10:50")
-@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_for_event.delay")
+@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_async")
 def test_product_deleted(mocked_webhook_trigger, settings, product):
     settings.PLUGINS = ["saleor.plugins.webhook.plugin.WebhookPlugin"]
     manager = get_plugins_manager()
@@ -252,51 +264,51 @@ def test_product_deleted(mocked_webhook_trigger, settings, product):
     assert variant_global_ids == expected_data_dict["variants"]
 
     mocked_webhook_trigger.assert_called_once_with(
-        WebhookEventAsyncType.PRODUCT_DELETED, expected_data
+        expected_data, WebhookEventAsyncType.PRODUCT_DELETED
     )
 
 
 @freeze_time("1914-06-28 10:50")
-@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_for_event.delay")
+@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_async")
 def test_product_variant_created(mocked_webhook_trigger, settings, variant):
     settings.PLUGINS = ["saleor.plugins.webhook.plugin.WebhookPlugin"]
     manager = get_plugins_manager()
     manager.product_variant_created(variant)
-
     expected_data = generate_product_variant_payload([variant])
+
     mocked_webhook_trigger.assert_called_once_with(
-        WebhookEventAsyncType.PRODUCT_VARIANT_CREATED, expected_data
+        expected_data, WebhookEventAsyncType.PRODUCT_VARIANT_CREATED
     )
 
 
 @freeze_time("1914-06-28 10:50")
-@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_for_event.delay")
+@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_async")
 def test_product_variant_updated(mocked_webhook_trigger, settings, variant):
     settings.PLUGINS = ["saleor.plugins.webhook.plugin.WebhookPlugin"]
     manager = get_plugins_manager()
     manager.product_variant_updated(variant)
-
     expected_data = generate_product_variant_payload([variant])
+
     mocked_webhook_trigger.assert_called_once_with(
-        WebhookEventAsyncType.PRODUCT_VARIANT_UPDATED, expected_data
+        expected_data, WebhookEventAsyncType.PRODUCT_VARIANT_UPDATED
     )
 
 
 @freeze_time("1914-06-28 10:50")
-@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_for_event.delay")
+@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_async")
 def test_product_variant_deleted(mocked_webhook_trigger, settings, variant):
     settings.PLUGINS = ["saleor.plugins.webhook.plugin.WebhookPlugin"]
     manager = get_plugins_manager()
     manager.product_variant_deleted(variant)
-
     expected_data = generate_product_variant_payload([variant])
+
     mocked_webhook_trigger.assert_called_once_with(
-        WebhookEventAsyncType.PRODUCT_VARIANT_DELETED, expected_data
+        expected_data, WebhookEventAsyncType.PRODUCT_VARIANT_DELETED
     )
 
 
 @freeze_time("1914-06-28 10:50")
-@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_for_event.delay")
+@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_async")
 def test_product_variant_out_of_stock(
     mocked_webhook_trigger, settings, variant_with_many_stocks
 ):
@@ -308,12 +320,12 @@ def test_product_variant_out_of_stock(
         [variant_with_many_stocks.stocks.first()]
     )
     mocked_webhook_trigger.assert_called_once_with(
-        WebhookEventAsyncType.PRODUCT_VARIANT_OUT_OF_STOCK, expected_data
+        expected_data, WebhookEventAsyncType.PRODUCT_VARIANT_OUT_OF_STOCK
     )
 
 
 @freeze_time("1914-06-28 10:50")
-@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_for_event.delay")
+@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_async")
 def test_product_variant_back_in_stock(
     mocked_webhook_trigger, settings, variant_with_many_stocks
 ):
@@ -325,90 +337,90 @@ def test_product_variant_back_in_stock(
         [variant_with_many_stocks.stocks.first()]
     )
     mocked_webhook_trigger.assert_called_once_with(
-        WebhookEventAsyncType.PRODUCT_VARIANT_BACK_IN_STOCK, expected_data
+        expected_data, WebhookEventAsyncType.PRODUCT_VARIANT_BACK_IN_STOCK
     )
 
 
 @freeze_time("1914-06-28 10:50")
-@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_for_event.delay")
+@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_async")
 def test_order_updated(mocked_webhook_trigger, settings, order_with_lines):
     settings.PLUGINS = ["saleor.plugins.webhook.plugin.WebhookPlugin"]
     manager = get_plugins_manager()
     manager.order_updated(order_with_lines)
-
     expected_data = generate_order_payload(order_with_lines)
+
     mocked_webhook_trigger.assert_called_once_with(
-        WebhookEventAsyncType.ORDER_UPDATED, expected_data
+        expected_data, WebhookEventAsyncType.ORDER_UPDATED
     )
 
 
 @freeze_time("1914-06-28 10:50")
-@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_for_event.delay")
+@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_async")
 def test_order_cancelled(mocked_webhook_trigger, settings, order_with_lines):
     settings.PLUGINS = ["saleor.plugins.webhook.plugin.WebhookPlugin"]
     manager = get_plugins_manager()
     manager.order_cancelled(order_with_lines)
-
     expected_data = generate_order_payload(order_with_lines)
+
     mocked_webhook_trigger.assert_called_once_with(
-        WebhookEventAsyncType.ORDER_CANCELLED, expected_data
+        expected_data, WebhookEventAsyncType.ORDER_CANCELLED
     )
 
 
 @freeze_time("1914-06-28 10:50")
-@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_for_event.delay")
+@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_async")
 def test_checkout_created(mocked_webhook_trigger, settings, checkout_with_items):
     settings.PLUGINS = ["saleor.plugins.webhook.plugin.WebhookPlugin"]
     manager = get_plugins_manager()
     manager.checkout_created(checkout_with_items)
-
     expected_data = generate_checkout_payload(checkout_with_items)
+
     mocked_webhook_trigger.assert_called_once_with(
-        WebhookEventAsyncType.CHECKOUT_CREATED, expected_data
+        expected_data, WebhookEventAsyncType.CHECKOUT_CREATED
     )
 
 
 @freeze_time("1914-06-28 10:50")
-@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_for_event.delay")
+@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_async")
 def test_checkout_updated(mocked_webhook_trigger, settings, checkout_with_items):
     settings.PLUGINS = ["saleor.plugins.webhook.plugin.WebhookPlugin"]
     manager = get_plugins_manager()
     manager.checkout_updated(checkout_with_items)
-
     expected_data = generate_checkout_payload(checkout_with_items)
+
     mocked_webhook_trigger.assert_called_once_with(
-        WebhookEventAsyncType.CHECKOUT_UPDATED, expected_data
+        expected_data, WebhookEventAsyncType.CHECKOUT_UPDATED
     )
 
 
 @freeze_time("1914-06-28 10:50")
-@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_for_event.delay")
+@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_async")
 def test_page_created(mocked_webhook_trigger, settings, page):
     settings.PLUGINS = ["saleor.plugins.webhook.plugin.WebhookPlugin"]
     manager = get_plugins_manager()
     manager.page_created(page)
-
     expected_data = generate_page_payload(page)
+
     mocked_webhook_trigger.assert_called_once_with(
-        WebhookEventAsyncType.PAGE_CREATED, expected_data
+        expected_data, WebhookEventAsyncType.PAGE_CREATED
     )
 
 
 @freeze_time("1914-06-28 10:50")
-@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_for_event.delay")
+@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_async")
 def test_page_updated(mocked_webhook_trigger, settings, page):
     settings.PLUGINS = ["saleor.plugins.webhook.plugin.WebhookPlugin"]
     manager = get_plugins_manager()
     manager.page_updated(page)
-
     expected_data = generate_page_payload(page)
+
     mocked_webhook_trigger.assert_called_once_with(
-        WebhookEventAsyncType.PAGE_UPDATED, expected_data
+        expected_data, WebhookEventAsyncType.PAGE_UPDATED
     )
 
 
 @freeze_time("1914-06-28 10:50")
-@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_for_event.delay")
+@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_async")
 def test_page_deleted(mocked_webhook_trigger, settings, page):
     settings.PLUGINS = ["saleor.plugins.webhook.plugin.WebhookPlugin"]
     manager = get_plugins_manager()
@@ -416,55 +428,57 @@ def test_page_deleted(mocked_webhook_trigger, settings, page):
     page.delete()
     page.id = page_id
     manager.page_deleted(page)
-
     expected_data = generate_page_payload(page)
 
     mocked_webhook_trigger.assert_called_once_with(
-        WebhookEventAsyncType.PAGE_DELETED, expected_data
+        expected_data, WebhookEventAsyncType.PAGE_DELETED
     )
 
 
 @freeze_time("1914-06-28 10:50")
-@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_for_event.delay")
+@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_async")
 def test_invoice_request(mocked_webhook_trigger, settings, fulfilled_order):
     settings.PLUGINS = ["saleor.plugins.webhook.plugin.WebhookPlugin"]
     manager = get_plugins_manager()
     invoice = fulfilled_order.invoices.first()
     manager.invoice_request(fulfilled_order, invoice, invoice.number)
     expected_data = generate_invoice_payload(invoice)
+
     mocked_webhook_trigger.assert_called_once_with(
-        WebhookEventAsyncType.INVOICE_REQUESTED, expected_data
+        expected_data, WebhookEventAsyncType.INVOICE_REQUESTED
     )
 
 
 @freeze_time("1914-06-28 10:50")
-@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_for_event.delay")
+@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_async")
 def test_invoice_delete(mocked_webhook_trigger, settings, fulfilled_order):
     settings.PLUGINS = ["saleor.plugins.webhook.plugin.WebhookPlugin"]
     manager = get_plugins_manager()
     invoice = fulfilled_order.invoices.first()
     manager.invoice_delete(invoice)
     expected_data = generate_invoice_payload(invoice)
+
     mocked_webhook_trigger.assert_called_once_with(
-        WebhookEventAsyncType.INVOICE_DELETED, expected_data
+        expected_data, WebhookEventAsyncType.INVOICE_DELETED
     )
 
 
 @freeze_time("1914-06-28 10:50")
-@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_for_event.delay")
+@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_async")
 def test_invoice_sent(mocked_webhook_trigger, settings, fulfilled_order):
     settings.PLUGINS = ["saleor.plugins.webhook.plugin.WebhookPlugin"]
     manager = get_plugins_manager()
     invoice = fulfilled_order.invoices.first()
     manager.invoice_sent(invoice, fulfilled_order.user.email)
     expected_data = generate_invoice_payload(invoice)
+
     mocked_webhook_trigger.assert_called_once_with(
-        WebhookEventAsyncType.INVOICE_SENT, expected_data
+        expected_data, WebhookEventAsyncType.INVOICE_SENT
     )
 
 
 @freeze_time("2020-03-18 12:00:00")
-@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_for_event.delay")
+@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_async")
 def test_notify_user(mocked_webhook_trigger, settings, customer_user, channel_USD):
     settings.PLUGINS = ["saleor.plugins.webhook.plugin.WebhookPlugin"]
     manager = get_plugins_manager(lambda: customer_user)
@@ -487,8 +501,7 @@ def test_notify_user(mocked_webhook_trigger, settings, customer_user, channel_US
         "channel_slug": channel_USD.slug,
         **get_site_context(),
     }
-
-    expected_data = {
+    data = {
         "notify_event": NotifyEventType.ACCOUNT_CONFIRMATION,
         "payload": payload,
         "meta": {
@@ -501,13 +514,47 @@ def test_notify_user(mocked_webhook_trigger, settings, customer_user, channel_US
         },
     }
     mocked_webhook_trigger.assert_called_once_with(
-        WebhookEventAsyncType.NOTIFY_USER,
-        json.dumps(expected_data, cls=CustomJsonEncoder),
+        json.dumps(data), WebhookEventAsyncType.NOTIFY_USER
     )
 
 
+def test_create_event_payload_reference_with_error(
+    webhook,
+    order_with_lines,
+    permission_manage_orders,
+    permission_manage_users,
+    permission_manage_products,
+    monkeypatch,
+):
+    mocked_client = MagicMock(spec=AsyncSQSConnection)
+    mocked_client_constructor = MagicMock(spec=boto3.client, return_value=mocked_client)
+
+    monkeypatch.setattr(
+        "saleor.plugins.webhook.tasks.boto3.client",
+        mocked_client_constructor,
+    )
+
+    webhook.app.permissions.add(permission_manage_orders)
+
+    webhook.target_url = "testy"
+    webhook.save()
+    expected_data = serialize("json", [order_with_lines])
+    event_payload = EventPayload.objects.create(payload=expected_data)
+    trigger_webhooks_async(event_payload, WebhookEventAsyncType.ORDER_CREATED)
+    delivery = EventDelivery.objects.first()
+    send_webhook_request_async(delivery.id)
+    attempt = EventDeliveryAttempt.objects.first()
+
+    assert delivery.webhook == webhook
+    assert delivery.event_type == WebhookEventAsyncType.ORDER_CREATED
+    assert attempt.response == "Unknown webhook scheme: ''"
+    assert delivery.status == "failed"
+    assert attempt.task_id == ANY
+    assert attempt.duration == ANY
+
+
 @freeze_time("1914-06-28 10:50")
-@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_for_event.delay")
+@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_async")
 def test_sale_created(mocked_webhook_trigger, settings, sale):
     settings.PLUGINS = ["saleor.plugins.webhook.plugin.WebhookPlugin"]
     manager = get_plugins_manager()
@@ -517,12 +564,12 @@ def test_sale_created(mocked_webhook_trigger, settings, sale):
     manager.sale_created(sale, current_catalogue=sale_catalogue_info)
     expected_data = generate_sale_payload(sale, current_catalogue=sale_catalogue_info)
     mocked_webhook_trigger.assert_called_once_with(
-        WebhookEventAsyncType.SALE_CREATED, expected_data
+        expected_data, WebhookEventAsyncType.SALE_CREATED
     )
 
 
 @freeze_time("1914-06-28 10:50")
-@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_for_event.delay")
+@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_async")
 def test_sale_updated(mocked_webhook_trigger, settings, sale, product_list):
     settings.PLUGINS = ["saleor.plugins.webhook.plugin.WebhookPlugin"]
     manager = get_plugins_manager()
@@ -545,12 +592,13 @@ def test_sale_updated(mocked_webhook_trigger, settings, sale, product_list):
         previous_catalogue=previous_sale_catalogue_info,
     )
     mocked_webhook_trigger.assert_called_once_with(
-        WebhookEventAsyncType.SALE_UPDATED, expected_data
+        expected_data,
+        WebhookEventAsyncType.SALE_UPDATED,
     )
 
 
 @freeze_time("1914-06-28 10:50")
-@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_for_event.delay")
+@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_async")
 def test_sale_deleted(mocked_webhook_trigger, settings, sale):
     settings.PLUGINS = ["saleor.plugins.webhook.plugin.WebhookPlugin"]
     manager = get_plugins_manager()
@@ -560,5 +608,57 @@ def test_sale_deleted(mocked_webhook_trigger, settings, sale):
     manager.sale_deleted(sale, previous_catalogue=sale_catalogue_info)
     expected_data = generate_sale_payload(sale, previous_catalogue=sale_catalogue_info)
     mocked_webhook_trigger.assert_called_once_with(
-        WebhookEventAsyncType.SALE_DELETED, expected_data
+        expected_data, WebhookEventAsyncType.SALE_DELETED
     )
+
+
+@mock.patch("saleor.plugins.webhook.plugin.send_webhook_request_async.delay")
+def test_event_delivery_retry(mocked_webhook_send, event_delivery, settings):
+    # given
+    settings.PLUGINS = ["saleor.plugins.webhook.plugin.WebhookPlugin"]
+    manager = get_plugins_manager()
+
+    # when
+    manager.event_delivery_retry(event_delivery)
+
+    # then
+    mocked_webhook_send.assert_called_once_with(event_delivery.pk)
+
+
+TEST_WEBHOOK_RESPONSE = WebhookResponse(
+    content="test_content",
+    request_headers={"headers": "test_request"},
+    response_headers={"headers": "test_response"},
+    duration=2.0,
+    status=EventDeliveryStatus.SUCCESS,
+)
+
+
+@mock.patch("saleor.plugins.webhook.tasks.clear_successful_delivery")
+@mock.patch(
+    "saleor.plugins.webhook.tasks.send_webhook_using_scheme_method",
+    return_value=TEST_WEBHOOK_RESPONSE,
+)
+def test_send_webhook_request_async(
+    mocked_send_response, mocked_clear_delivery, event_delivery
+):
+    send_webhook_request_async(event_delivery.pk)
+
+    mocked_send_response.assert_called_once_with(
+        event_delivery.webhook.target_url,
+        "mirumee.com",
+        event_delivery.webhook.secret_key,
+        event_delivery.event_type,
+        event_delivery.payload.payload,
+    )
+    mocked_clear_delivery.assert_called_once_with(event_delivery)
+    attempt = EventDeliveryAttempt.objects.filter(delivery=event_delivery).first()
+    delivery = EventDelivery.objects.get(id=event_delivery.pk)
+    assert attempt.status == EventDeliveryStatus.SUCCESS
+    assert attempt.response == TEST_WEBHOOK_RESPONSE.content
+    assert attempt.response_headers == json.dumps(
+        TEST_WEBHOOK_RESPONSE.response_headers
+    )
+    assert attempt.request_headers == json.dumps(TEST_WEBHOOK_RESPONSE.request_headers)
+    assert attempt.duration == TEST_WEBHOOK_RESPONSE.duration
+    assert delivery.status == EventDeliveryStatus.SUCCESS